refactor: artists folders
