+++ app/modules/artists/resolvers/artists.js
