+++ app/modules/artists/schemas/artists.graphql
