+++ app/modules/artists/services/artists.js
